feat: refactor OtakiageFeature and comment out useApproval in Connect
