+++ pages/otakiage/index.tsx
@@ -30,6 +30,7 @@ import {
   useSendAllOmamoriWithMx
 } from '@/hooks/useOtakiageContract'
 import { IS_RELEASED, IS_EVENT_DAY } from '@/constants/Otakiage'
+import OtakiageFeature from '@/components/OtakiageFeature'
 
 const OtakiagePage: NextPage = () => {
   const isMounted = useMounted()
@@ -113,150 +114,7 @@ const OtakiagePage: NextPage = () => {
 
   return (
     <Layout>
-      {isMounted && (
-        <Box>
-          <Flex direction="column" alignItems="center" mt={10}>
-            <Heading
-              className="text_serif"
-              size="2xl"
-              mb={10}
-              textAlign="center"
-              width="100%"
-            >
-              HENKAKU
-              <br />「{ot('OTAKIAGE')}」
-            </Heading>
-          </Flex>
-          <Flex
-            marginInline="calc((100vw - 100%) * -0.5)"
-            justifyContent="center"
-            backgroundColor="#1A202C"
-          >
-            <Box
-              position="relative"
-              width="100%"
-              maxWidth="768px"
-              textAlign="center"
-            >
-              <Image
-                src="/torii_gate.png"
-                alt=""
-                sx={{
-                  marginTop: '-40px',
-                  marginBottom: '-30px',
-                  filter: `invert(100%) brightness(70%)`
-                }}
-              />
-              <Flex
-                width="100%"
-                height="100%"
-                position="absolute"
-                left="0"
-                top="0"
-              >
-                <Orb />
-                <Orb color="#f00" highlight="#f90" delay={0.5} />
-                <Orb color="#ff0" highlight="#ff9" delay={1} />
-                <Orb color="#0f0" highlight="#0f9" delay={1.5} />
-                <Orb color="#0ff" highlight="#9ff" delay={2} />
-              </Flex>
-            </Box>
-          </Flex>
-          {/* <Text className="text_serif" fontSize="lg" mt={10}>
-            <NewlineToBr>{ot('OMAMORI_EXPLANATION')}</NewlineToBr>
-          </Text> */}
-          {address && (
-            <>
-              {!isOtakiaged && (
-                <>
-                  <Heading className="text_serif" size="xl" mt={10} pb={10}>
-                    {ot('USER_OMAMORI_LIST')}
-                  </Heading>
-                  <Box mt={0} pb={20}>
-                    {<OmamoriesList items={userMintedOmamories} />}
-                  </Box>
-                  {(IS_RELEASED ||
-                    isAdmin ||
-                    !(process.env.NODE_ENV === 'production')) && (
-                    <>
-                      <Flex
-                        justifyContent="center"
-                        alignItems="center"
-                        direction="column"
-                        mt={10}
-                        pb={100}
-                        maxWidth="240px"
-                        margin="0 auto"
-                        gap={10}
-                      >
-                        {userMintedOmamories &&
-                          userMintedOmamories.length > 0 && (
-                            <>
-                              {!isApproved && !isSuccessApprove ? (
-                                <Button
-                                  w="100%"
-                                  loadingText="approving..."
-                                  isLoading={isApproving}
-                                  onClick={approve}
-                                >
-                                  {ot('JOIN')}
-                                </Button>
-                              ) : (
-                                <>
-                                  {userHoldingOmamories.length > 0 &&
-                                  !isSuccessSendAllOmamori ? (
-                                    <Button
-                                      w="100%"
-                                      loadingText="sending..."
-                                      isLoading={isSending}
-                                      onClick={sendAllOmamori}
-                                    >
-                                      {ot('SEND_OMAMORI')}
-                                    </Button>
-                                  ) : (
-                                    <Box w="100%">
-                                      <Text>「御守り」を送りました！</Text>
-                                      <Text>
-                                        お焚き上げイベントまで、お待ちください！
-                                      </Text>
-                                    </Box>
-                                  )}
-                                </>
-                              )}
-                            </>
-                          )}
-                        {IS_EVENT_DAY && isAdmin && (
-                          <Button
-                            w="100%"
-                            loadingText="otakiage..."
-                            isLoading={isLoadingOtakiage}
-                            onClick={otakiage}
-                          >
-                            {ot('OTAKIAGE')}
-                          </Button>
-                        )}
-                      </Flex>
-                    </>
-                  )}
-                </>
-              )}
-              {(isOtakiaged || isSuccessOtakiage) &&
-                (IS_RELEASED ||
-                  isAdmin ||
-                  !(process.env.NODE_ENV === 'production')) && (
-                  <Box pb={100}>
-                    {IS_EVENT_DAY && (
-                      <Text textAlign="center">お焚き上げ！！！！</Text>
-                    )}
-                    {!IS_EVENT_DAY && address && (
-                      <Text textAlign="center">{ot('END_OTAKIAGE')}</Text>
-                    )}
-                  </Box>
-                )}
-            </>
-          )}
-        </Box>
-      )}
+      <OtakiageFeature />
     </Layout>
   )
 }
